docker/api_flask_rag/requirements.txt
Flask>=2.2
flask[async]
requests==2.31.0
pydantic
litellm
langchain_community
pypdf
fastembed
chromadb
langchain-chroma